refactor : LottoGeneratorTest 객체 동등성비교 통한 테스트
