+++ src/test/java/domain/LottoGeneratorTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.IntStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,8 +12,9 @@ class LottoGeneratorTest {
     @Test
     void generateLotto() {
         List<Integer> numbers = Arrays.asList(1, 2, 3, 4, 5, 6);
-        Lotto lotto = LottoGenerator.generateLotto(numbers);
+        Lotto lotto = new Lotto(numbers);
+        Lotto target = LottoGenerator.generateLotto(numbers);
 
-        assertThat(lotto.getNumbers()).isEqualTo(Arrays.asList(1, 2, 3, 4, 5, 6));
+        assertThat(target.equals(lotto)).isTrue();
     }
 }
