+++ src/test/java/domain/RandomNumberGeneratorTest.java
@@ -1,6 +1,5 @@
-package util;
+package domain;
 
-import domain.RandomNumberGenerator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
